FEAT -> Adc nota final na rating -> mudando nota final de group para double

+++ hackweek-backend/Models/GroupModel.cs
@@ -9,7 +9,7 @@
         public PropositionModel? Proposition { get; set; }
         public string ProjectName { get; set; } = string.Empty;
         public string ProjectDescription { get; set; } = string.Empty;
-        public uint FinalGrade { get; set; }
+        public double FinalGrade { get; set; }
         public uint Position { get; set; }
         public DateTime? StartTime { get; set; }
         public DateTime? EndTime { get; set; }
